requirements.txt
fastapi[standard]>=0.115.2
pillow>=10.4.0